Split monolithic Init.hs module into 5 sub-modules.
Init.Command - Main cabal init logic that ties everything together.
Init.Defaults - Default values for items not specified in config or arguments
Init.FileCreators - All the createXyz functions which write files (LICENSE, Main.hs, etc.)
Init.Prompt - Utilities for prompting the user for selections
Init.Utils - Misc utilities shared across the above modules
