+++ cabal-install/cabal-install.cabal
@@ -204,9 +204,14 @@ executable cabal
         Distribution.Client.IndexUtils.IndexState
         Distribution.Client.IndexUtils.Timestamp
         Distribution.Client.Init
+        Distribution.Client.Init.Command
+        Distribution.Client.Init.Defaults
+        Distribution.Client.Init.FileCreators
         Distribution.Client.Init.Heuristics
         Distribution.Client.Init.Licenses
+        Distribution.Client.Init.Prompt
         Distribution.Client.Init.Types
+        Distribution.Client.Init.Utils
         Distribution.Client.Install
         Distribution.Client.InstallPlan
         Distribution.Client.InstallSymlink
